Move Button and Input into components/common/ directory, make index.js there and export using * so that you may export those reusable components using { } destructuring; update all components to use that new import type from ./common
