+++ components/common/Button.js
@@ -30,4 +30,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export default Button;
+export { Button };
