+++ components/common/Input.js
@@ -39,4 +39,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export default Input;
+export { Input };
